Further changes with the order summary
