+++ src/Web/Grand.Web/Controllers/CheckoutController.cs
@@ -370,7 +370,7 @@ namespace Grand.Web.Controllers
             await _userFieldService.SaveField<string>(_workContext.CurrentCustomer,
                 SystemCustomerFieldNames.SelectedPaymentMethod, null, _workContext.CurrentStore.Id);
 
-            var confirmOrderModel = await _mediator.Send(new GetConfirmOrder() { Cart = cart, Customer = _workContext.CurrentCustomer });
+            var confirmOrderModel = await _mediator.Send(new GetConfirmOrder() { Cart = cart, Customer = _workContext.CurrentCustomer, Language = _workContext.WorkingLanguage, Store = _workContext.CurrentStore });
             return Json(new
             {
                 update_section = new UpdateSectionJsonModel {
@@ -397,7 +397,7 @@ namespace Grand.Web.Controllers
                     (paymentMethod.PaymentMethodType == PaymentMethodType.Redirection
                     && _paymentSettings.SkipPaymentInfo))
             {
-                var confirmOrderModel = await _mediator.Send(new GetConfirmOrder() { Cart = cart, Customer = _workContext.CurrentCustomer });
+                var confirmOrderModel = await _mediator.Send(new GetConfirmOrder() { Cart = cart, Customer = _workContext.CurrentCustomer, Language = _workContext.WorkingLanguage, Store = _workContext.CurrentStore });
                 return Json(new
                 {
                     update_section = new UpdateSectionJsonModel {
@@ -854,7 +854,7 @@ namespace Grand.Web.Controllers
                     await _userFieldService.SaveField<string>(_workContext.CurrentCustomer,
                         SystemCustomerFieldNames.SelectedPaymentMethod, null, _workContext.CurrentStore.Id);
 
-                    var confirmOrderModel = await _mediator.Send(new GetConfirmOrder() { Cart = cart, Customer = _workContext.CurrentCustomer });
+                    var confirmOrderModel = await _mediator.Send(new GetConfirmOrder() { Cart = cart, Customer = _workContext.CurrentCustomer, Language = _workContext.WorkingLanguage, Store = _workContext.CurrentStore });
                     return Json(new
                     {
                         update_section = new UpdateSectionJsonModel {
@@ -923,7 +923,7 @@ namespace Grand.Web.Controllers
                             SystemCustomerFieldNames.PaymentTransaction, paymentTransaction.Id, _workContext.CurrentStore.Id);
                     }
 
-                    var confirmOrderModel = await _mediator.Send(new GetConfirmOrder() { Cart = cart, Customer = _workContext.CurrentCustomer });
+                    var confirmOrderModel = await _mediator.Send(new GetConfirmOrder() { Cart = cart, Customer = _workContext.CurrentCustomer, Language = _workContext.WorkingLanguage, Store = _workContext.CurrentStore });
                     return Json(new
                     {
                         update_section = new UpdateSectionJsonModel {
