+++ src/Web/Grand.Web/Features/Models/Checkout/GetConfirmOrder.cs
@@ -1,5 +1,7 @@
 ﻿using Grand.Domain.Customers;
+using Grand.Domain.Localization;
 using Grand.Domain.Orders;
+using Grand.Domain.Stores;
 using Grand.Web.Models.Checkout;
 using MediatR;
 
@@ -8,6 +10,8 @@ namespace Grand.Web.Features.Models.Checkout
     public class GetConfirmOrder : IRequest<CheckoutConfirmModel>
     {
         public Customer Customer { get; set; }
+        public Language Language { get; set; }
+        public Store Store { get; set; }
         public IList<ShoppingCartItem> Cart { get; set; }
     }
 }
