+++ src/Web/Grand.Web/Views/ShoppingCart/_ModelScript.cshtml
@@ -32,6 +32,7 @@
                 body.classList.add("cart-view");
             },
             updateCart() {
+                alert('updatecart');
                 axios({
                     baseURL: '@Url.RouteUrl("ShoppingCartSummary")',
                     method: 'get',
